+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Version {next}
+  - Allow new notes to be made by a JSON request from a micropub client
+
 ## Version 0.0.7.1 (2016-07-04)
   - Minor style fixes