+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Version 0.0.8 (2016-07-13)
+  - Allow new notes to be made by a JSON request from a micropub client
+  - Add DependencyCI support
+
 ## Version 0.0.7.1 (2016-07-04)
   - Minor style fixes